third_party/blink/web_tests/external/wpt/css/css-align/blocks/justify-self-vrl-rtl-htb-expected.txt

This is a testharness.js-based test.
[FAIL] .item 4
  assert_equals: \n<div class="container">\n  <div class="item" style="justify-self: self-start;" data-expected-height="20" data-offset-y="0"></div>\n</div>\noffsetTop expected 0 but got 20
[FAIL] .item 12
  assert_equals: \n<div class="container">\n  <div class="item rtl" style="justify-self: self-start;" data-expected-height="20" data-offset-y="0"></div>\n</div>\noffsetTop expected 0 but got 20
Harness: the test ran to completion.
